Sorta fixed the periodic table, made it so the atomic weight would stop showing if the screen was too smalll
The problem is that when the atomic weight stops showing, the height of the element decreases? There has been this weird empty space on the flexbox that has been bugging me and i think it has something to do with the how the height of the flexbox is calculated first and then the size of the text (for atomic number) is shrunk after, so the leftover space is just left there, which makes the box become rectangular.

Maybe if I put the aspect ratio thing after the children of the button have their heights calculated, the new aspect ratio will be calculated and that empty space will go away?